+++ R/kwic.R
@@ -58,14 +58,14 @@ setMethod("show", "kwic", function(object){
 })
 
 
-#' @details The \code{knit_print} method will be called by knitr to render
-#'   `kwic` objects as a DataTable  \code{htmlwidget} when rendering a R
+#' @details The `knit_print()` method will be called by knitr to render
+#'   `kwic` objects as a DataTable  `htmlwidget` when rendering a R
 #'   Markdown document as html. It will usually be necessary to explicitly state
 #'   "render = knit_print" in the chunk options. The option
 #'   `polmineR.pagelength` controls the number of lines displayed in the
 #'   resulting `htmlwidget`. Note that including htmlwidgets in html documents
 #'   requires that pandoc is installed. To avoid an error, a formatted
-#'   \code{data.table} is returned by \code{knit_print} if pandoc is not
+#'   `data.table` is returned by `knit_print()` if pandoc is not
 #'   available.
 #' @importFrom knitr knit_print
 #' @exportMethod knit_print